projects/python-sensor-gateway/requirements.txt
bleak==0.20.2
pydantic==1.10.9
python-dotenv==1.0.0